switch active app to the child health one
